refactor: error/success code 리팩토링
-enum명과 code통일
-가독성 있게 이름 변경
-errorcode 세분화
-도메인별 exception파일 삭제
