+++ src/main/java/com/connecteamed/server/domain/member/code/MemberSuccessCode.java
@@ -1,4 +1,4 @@
-package com.connecteamed.server.domain.member.exception.code;
+package com.connecteamed.server.domain.member.code;
 
 import com.connecteamed.server.global.apiPayload.code.BaseSuccessCode;
 import lombok.AllArgsConstructor;
@@ -10,9 +10,7 @@ import org.springframework.http.HttpStatus;
 @AllArgsConstructor
 public enum MemberSuccessCode implements BaseSuccessCode {
 
-    FOUND(HttpStatus.OK,
-            "MEMBER200_1",
-            "성공적으로 사용자를 조회했습니다."),
+    FOUND(HttpStatus.OK, "MEMBER_FOUND", "성공적으로 사용자를 조회했습니다."),
     ;
 
     private final HttpStatus status;
