--- src/main/java/com/connecteamed/server/domain/member/exception/MemberException.java
@@ -1,10 +1,0 @@
-package com.connecteamed.server.domain.member.exception;
-
-import com.connecteamed.server.global.apiPayload.code.BaseErrorCode;
-import com.connecteamed.server.global.apiPayload.exception.GeneralException;
-
-public class MemberException extends GeneralException {
-    public MemberException(BaseErrorCode code) {
-        super(code);
-    }
-}
