--- src/main/java/com/connecteamed/server/domain/member/exception/code/MemberErrorCode.java
@@ -1,22 +1,0 @@
-package com.connecteamed.server.domain.member.exception.code;
-
-import com.connecteamed.server.global.apiPayload.code.BaseErrorCode;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@AllArgsConstructor
-public enum MemberErrorCode implements BaseErrorCode {
-
-    NOT_FOUND(HttpStatus.NOT_FOUND,
-            "MEMBER404_1",
-            "해당 사용자를 찾지 못했습니다."),
-    ID_DUPLICATED(HttpStatus.BAD_REQUEST, "ALREADY_EXIST","이미 존재하는 ID입니다")
-    ;
-
-
-    private final HttpStatus status;
-    private final String code;
-    private final String message;
-}
