+++ src/main/java/com/connecteamed/server/global/auth/CustomUserDetailsService.java
@@ -1,9 +1,9 @@
 package com.connecteamed.server.global.auth;
 
 
-import com.connecteamed.server.domain.member.exception.MemberException;
-import com.connecteamed.server.domain.member.exception.code.MemberErrorCode;
+import com.connecteamed.server.domain.member.code.MemberErrorCode;
 import com.connecteamed.server.domain.member.repository.MemberRepository;
+import com.connecteamed.server.global.apiPayload.exception.GeneralException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
@@ -15,9 +15,9 @@ public class CustomUserDetailsService implements UserDetailsService {
     private final MemberRepository memberRepository;
 
     @Override
-    public CustomUserDetails loadUserByUsername(String loginId) throws MemberException {
+    public CustomUserDetails loadUserByUsername(String loginId) throws GeneralException {
         return memberRepository.findByLoginId(loginId)
                 .map(CustomUserDetails::new)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.NOT_FOUND));
+                .orElseThrow(() -> new GeneralException(MemberErrorCode.MEMBER_NOT_FOUND));
     }
 }
