+++ src/main/java/com/connecteamed/server/global/auth/exception/code/AuthErrorCode.java
@@ -10,15 +10,15 @@ import org.springframework.http.HttpStatus;
 public enum AuthErrorCode implements BaseErrorCode {
 
 
-    INVALID_PASSWORD(HttpStatus.UNAUTHORIZED,"MEMBER4011", "비밀번호가 틀렸습니다"),
-    TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "AUTH4013", "토큰이 없거나 만료되었습니다."),
-    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "AUTH4014", "유효하지 않은 토큰입니다."),
-    INVALID_REFRESH_TOKEN(HttpStatus.UNAUTHORIZED,"INVALID_ACCESS_TOKEN","유효하지 않은 refresh 토큰입니다"),
+    INCORRECT_PASSWORD(HttpStatus.UNAUTHORIZED,"INCORRECT_PASSWORD", "비밀번호가 틀렸습니다"),
+    TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "TOKEN_EXPIRED", "토큰이 만료되었습니다."),
+    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "INVALID_TOKEN", "유효하지 않은 토큰입니다."),
+    INVALID_REFRESH_TOKEN(HttpStatus.UNAUTHORIZED,"INVALID_REFRESH_TOKEN","유효하지 않은 리프레시 토큰입니다"),
     REFRESH_TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED,"REFRES_TOKEN_EXPIRED","리프레시 토큰이 만료되었습니다. 다시 로그인해 주세요")
     ,
     // 회원가입 시 중복된 아이디가 있을 경우
-    DUPLICATE_LOGIN_ID(HttpStatus.CONFLICT, "AUTH4091", "이미 존재하는 아이디입니다."),
-    EMPTY_AUTHENTICATION(HttpStatus.UNAUTHORIZED, "AUTH4011", "인증 정보가 존재하지 않습니다.")
+    DUPLICATE_LOGIN_ID(HttpStatus.CONFLICT, "DUPLICATE_LOGIN_ID", "이미 존재하는 아이디입니다."),
+    EMPTY_AUTHENTICATION(HttpStatus.UNAUTHORIZED, "EMPTY_AUTHENTICATION", "인증 정보가 존재하지 않습니다.")
     ;
 
     private final HttpStatus status;
